re #1658: Updated per review comments * Created a separate DelegatingQueryLogic * Created separate predicates for auths and query parameters that can   be reused elsewhere
