+++ warehouse/query-core/src/main/java/datawave/query/tables/remote/RemoteQueryLogic.java
@@ -1,7 +1,11 @@
 package datawave.query.tables.remote;
 
 import datawave.webservice.common.remote.RemoteQueryService;
+import datawave.webservice.query.logic.QueryLogic;
 
-public interface RemoteQueryLogic {
+/**
+ * A remote query logic is is a query logic that uses a remote query service.
+ */
+public interface RemoteQueryLogic<T> extends QueryLogic<T> {
     public void setRemoteQueryService(RemoteQueryService service);
 }
