+++ web-services/query/src/main/java/datawave/webservice/query/logic/filtered/QueryLogicFilterByAuth.java
@@ -1,27 +1,18 @@
 package datawave.webservice.query.logic.filtered;
 
 import datawave.webservice.query.Query;
+import datawave.webservice.query.predicate.ProxiedAuthorizationsPredicate;
 import org.apache.accumulo.core.security.Authorizations;
-import org.apache.accumulo.core.security.ColumnVisibility;
-import org.apache.accumulo.core.security.VisibilityEvaluator;
-import org.apache.accumulo.core.security.VisibilityParseException;
 
 import java.util.Set;
+import java.util.stream.Collectors;
 
-public class QueryLogicFilterByAuth implements FilteredQueryLogic.QueryLogicFilter {
-    
-    public static enum MatchType {
-        // match only the first of the Authorizations (presumably the user)
-        FIRST,
-        // match all of the Authorizations (the user and proxies)
-        ALL,
-    }
-    
-    // A visibility string to be matched against the auths being used for the query
-    private ColumnVisibility visibility;
-    
-    // Default is to only require the user visibilities to match
-    private MatchType matchType = MatchType.ALL;
+/**
+ * This is a filter for the FilteredQueryLogic that will run the delegate query logic if the auths requested match a specified visibility (as defined by
+ * accumulo's ColumnVisibility). In addition to the visibility, one can specify that only the first proxied user is matched (presumably the user), and one can
+ * negate the matching of the visibility.
+ */
+public class QueryLogicFilterByAuth extends ProxiedAuthorizationsPredicate implements FilteredQueryLogic.QueryLogicFilter {
     
     // if negated than the negation of the match is returned
     private boolean negated = false;
@@ -44,51 +35,11 @@ public class QueryLogicFilterByAuth implements FilteredQueryLogic.QueryLogicFilt
     
     @Override
     public boolean canRunQuery(Query settings, Set<Authorizations> auths) {
-        boolean canRunQuery = matches(auths);
+        boolean canRunQuery = test(auths.stream().collect(Collectors.toList()));
         if (negated) {
             canRunQuery = !canRunQuery;
         }
         return canRunQuery;
-    }
-    
-    private boolean matches(Set<Authorizations> auths) {
-        // match the visibility against the auths. If any of the authorizations to not match
-        ColumnVisibility vis = getVisibility();
-        for (Authorizations auth : auths) {
-            VisibilityEvaluator ve = new VisibilityEvaluator(auth);
-            try {
-                if (ve.evaluate(vis)) {
-                    if (matchType == MatchType.FIRST) {
-                        return true;
-                    }
-                } else {
-                    return false;
-                }
-            } catch (VisibilityParseException e) {
-                throw new RuntimeException(e);
-            }
-        }
-        return true;
-    }
-    
-    public ColumnVisibility getVisibility() {
-        return visibility;
-    }
-    
-    public void setVisibility(ColumnVisibility visibility) {
-        this.visibility = visibility;
-    }
-    
-    public void setVisibility(String visibility) {
-        setVisibility(new ColumnVisibility(visibility));
-    }
-    
-    public MatchType getMatchType() {
-        return matchType;
-    }
-    
-    public void setMatchType(MatchType matchType) {
-        this.matchType = matchType;
     }
     
     public boolean isNegated() {
